2023.04.27 added minimap. components now need to be repaired first

+++ TI3/Assets/Scripts/Data/GridObjectDeliveryData.cs
@@ -11,5 +11,5 @@ public class GridObjectDeliveryData
     public int desiredLength;
     public int snapValue;
     public int maximumIntegrityPoints;
-    public int currentIntegrityPoints;
+    public float currentIntegrityPoints;
 }
